refactor code to add support for different databases
